Implement Op::Label, Op::JmpLabel, Op::JmpIfNotLabel
For this implementation I implemented a Label structure
to use with the relocation infrastructure.
There was a problem, and that is that there was a need to
have the `label` only with the function local "context"
(needs to know which function defined the label)
because of how the `relocation` system is implemented.

One way to avoid this, would be to relocate each label at
the end of function generation, but I am not sure if it would
be a good solution.