[FIX] Introduce "trampolines" to SH4 litterals
Instead of being located at the end of a function, literals now litter
your code to be as close as possible.

This also now brings the buildable tests to pretty much every single one
which aren't designed to be platform-specific!

I just need a test SH4 runner, so if anyone wants to contribute, feel
free.